Update display_make_Test.py
Size input error retry

Added on line90 to 105: A list in a while loop that helps to check if the inputted characters are integers. If they are then it is flagged to end the loop if not the loop continues. It checks if each individual character is inside the list if a character is good.